Moved classes Assets files to models
